Move approve function to figure input
