+++ src/views/EventReview/index.tsx
@@ -1,6 +1,6 @@
-import React, { useCallback, useContext, useMemo, useState } from 'react';
+import React, { useMemo, useState } from 'react';
 import { useParams } from 'react-router-dom';
-import { gql, useMutation, useQuery } from '@apollo/client';
+import { useQuery } from '@apollo/client';
 import { removeNull } from '@togglecorp/toggle-form';
 import {
     isDefined,
@@ -14,8 +14,6 @@ import {
     FIGURE_OPTIONS,
 } from '#components/forms/EntryForm/queries';
 import {
-    ApproveFigureMutation,
-    ApproveFigureMutationVariables,
     FigureListQuery,
     FigureListQueryVariables,
     FigureOptionsForEntryFormQuery,
@@ -41,23 +39,8 @@ import { ViolenceContextOption } from '#components/selections/ViolenceContextMul
 import { OrganizationOption } from '#components/selections/OrganizationSelectInput';
 import Preview from '#components/Preview';
 import FigureInput from '#components/forms/EntryForm/FigureInput';
-import NotificationContext from '#components/NotificationContext';
 
 import styles from './styles.css';
-
-const APPROVE_FIGURE = gql`
-    mutation ApproveFigure($id: ID!) {
-        approveFigure(id: $id) {
-            errors
-            ok
-            result {
-                id
-                reviewStatus
-                reviewStatusDisplay
-            }
-        }
-    }
-`;
 
 const mode = 'view';
 const trafficLightShown = mode === 'view';
@@ -84,10 +67,6 @@ interface Props {
 }
 
 function EventReview(props: Props) {
-    const {
-        notify,
-        notifyGQLError,
-    } = useContext(NotificationContext);
     const {
         data: figureOptionsData,
         loading: figureOptionsLoading,
@@ -158,48 +137,6 @@ function EventReview(props: Props) {
         },
     });
 
-    const [
-        approveFigure,
-        { loading: approvingFigure },
-    ] = useMutation<ApproveFigureMutation, ApproveFigureMutationVariables>(
-        APPROVE_FIGURE,
-        {
-            onCompleted: (response) => {
-                const { approveFigure: approveResponse } = response;
-                if (!approveResponse) {
-                    return;
-                }
-                const { errors, result } = approveResponse;
-                if (errors) {
-                    notifyGQLError(errors);
-                }
-                if (result) {
-                    notify({
-                        children: 'Figure approved successfully!',
-                        variant: 'success',
-                    });
-                }
-            },
-            onError: (error) => {
-                notify({
-                    children: error.message,
-                    variant: 'error',
-                });
-            },
-        },
-    );
-
-    const handleFigureApprove = useCallback(
-        (id: string) => {
-            approveFigure({
-                variables: {
-                    id,
-                },
-            });
-        },
-        [approveFigure],
-    );
-
     const {
         preview,
         attachment,
@@ -256,7 +193,7 @@ function EventReview(props: Props) {
                             onChange={() => null}
                             onRemove={() => null}
                             error={undefined}
-                            disabled={getFiguresLoading || approvingFigure}
+                            disabled={getFiguresLoading}
                             mode={mode}
                             optionsDisabled={!!figureOptionsError || !!figureOptionsLoading}
                             tagOptions={tagOptions}
@@ -296,7 +233,6 @@ function EventReview(props: Props) {
                             trafficLightShown={trafficLightShown}
                             organizations={organizations}
                             setOrganizations={setOrganizations}
-                            onFigureApprove={handleFigureApprove}
                         />
                     ))}
                 </Container>
